fix: resolve JWT auth issue

+++ be-shoestore/src/main/java/com/shoestore/Server/service/impl/CartItemServiceImpl.java
@@ -1,16 +1,25 @@
 package com.shoestore.Server.service.impl;
 
+import com.shoestore.Server.dto.request.CartItemDTO;
+import com.shoestore.Server.dto.response.CartItemResponse;
+import com.shoestore.Server.dto.response.PaginationResponse;
 import com.shoestore.Server.entities.Cart;
 import com.shoestore.Server.entities.CartItem;
 import com.shoestore.Server.entities.CartItemKey;
 import com.shoestore.Server.entities.ProductDetail;
+import com.shoestore.Server.mapper.CartItemMapper;
+import com.shoestore.Server.mapper.ProductDetailMapper;
+import com.shoestore.Server.mapper.ProductMapper;
 import com.shoestore.Server.repositories.CartItemRepository;
 import com.shoestore.Server.repositories.CartRepository;
 import com.shoestore.Server.repositories.ProductDetailRepository;
 import com.shoestore.Server.service.CartItemService;
+import com.shoestore.Server.service.PromotionService;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
+import com.shoestore.Server.service.PaginationService;
 import java.util.Optional;
 
 @Service
@@ -18,19 +27,41 @@ public class CartItemServiceImpl implements CartItemService {
     private final ProductDetailRepository productDetailRepository;
     private final CartItemRepository cartItemRepository;
     private final CartRepository cartRepository;
-
-    public CartItemServiceImpl(ProductDetailRepository productDetailRepository, CartItemRepository cartItemRepository, CartRepository cartRepository) {
+    private final PaginationService paginationService ;
+    private final CartItemMapper cartItemMapper;
+    private final ProductDetailMapper productDetailMapper;
+    private final PromotionService promotionService;
+    private final ProductMapper productMapper;
+    public CartItemServiceImpl(ProductDetailRepository productDetailRepository, CartItemRepository cartItemRepository, CartRepository cartRepository, PaginationService paginationService, CartItemMapper cartItemMapper, ProductDetailMapper productDetailMapper, PromotionService promotionService, ProductMapper productMapper) {
         this.productDetailRepository = productDetailRepository;
         this.cartItemRepository = cartItemRepository;
         this.cartRepository = cartRepository;
+        this.paginationService = paginationService;
+        this.cartItemMapper = cartItemMapper;
+        this.productDetailMapper = productDetailMapper;
+        this.promotionService = promotionService;
+        this.productMapper = productMapper;
+    }
+    @Override
+    public PaginationResponse<CartItemResponse> getCartItemsByCartId(int cartId, int page, int pageSize) {
+        Pageable pageable = paginationService.createPageable(page, pageSize);
+
+        Page<CartItem> cartItemsPage = cartItemRepository.findCartItemsByCartId(cartId, pageable);
+
+        Page<CartItemResponse> cartItemResponses = cartItemsPage.map(cartItem -> {
+            ProductDetail productDetail = cartItem.getProductDetail();
+            CartItemResponse cartItemResponse = new CartItemResponse();
+            cartItemResponse.setCartItemDTO(cartItemMapper.toCartItemDTO(cartItem));
+            cartItemResponse.setProductName(productDetail.getProduct().getProductName());
+            cartItemResponse.setProductDetailDTO(productDetailMapper.toDto(productDetail));
+            cartItemResponse.setProductPrice(promotionService.getDiscountedPrice(productMapper.toDto(productDetail.getProduct())));
+            return cartItemResponse;
+        });
+        return paginationService.paginate(cartItemResponses);
     }
 
     @Override
-    public List<CartItem> getCartItemsByCartId(int cartId) {
-        return cartItemRepository.findCartItemsByCartId(cartId);
-    }
-
-    public CartItem addCartItem(CartItem cartItem) {
+    public CartItemDTO addCartItem(CartItem cartItem) {
         Cart cart = cartRepository.findById(cartItem.getId().getCartId())
                 .orElseThrow(() -> new IllegalArgumentException("Cart not found"));
         ProductDetail productDetail = productDetailRepository.findById(cartItem.getId().getProductDetailId())
@@ -41,43 +72,41 @@ public class CartItemServiceImpl implements CartItemService {
 
         cartItem.setCart(cart);
         cartItem.setProductDetail(productDetail);
-        return cartItemRepository.save(cartItem);
+        CartItem savedCartItem = cartItemRepository.save(cartItem);
+
+        return cartItemMapper.toCartItemDTO(savedCartItem);
     }
 
     @Override
-    public CartItem getCartItemById(CartItemKey cartItemKey) {
-        return cartItemRepository.findById(cartItemKey).orElse(null);
+    public CartItemDTO getCartItemById(CartItemKey cartItemKey) {
+        CartItem cartItem = cartItemRepository.findById(cartItemKey).orElse(null);
+        return cartItem != null ? cartItemMapper.toCartItemDTO(cartItem) : null;
     }
 
     @Override
-    public CartItem updateQuantity(CartItemKey id, CartItem cartItem) {
-        Optional<CartItem> existCartItem= cartItemRepository.findById(id);
-        if(existCartItem.isPresent()){
-            CartItem entityCartItem=existCartItem.get();
+    public CartItemDTO updateQuantity(CartItemKey id, CartItem cartItem) {
+        Optional<CartItem> existCartItem = cartItemRepository.findById(id);
+        if (existCartItem.isPresent()) {
+            CartItem entityCartItem = existCartItem.get();
 
             entityCartItem.setQuantity(cartItem.getQuantity());
             entityCartItem.setCart(cartItem.getCart());
             entityCartItem.setProductDetail(cartItem.getProductDetail());
             entityCartItem.setSubTotal(cartItem.getSubTotal());
-            return cartItemRepository.save(entityCartItem);
+
+            CartItem savedCartItem = cartItemRepository.save(entityCartItem);
+            return cartItemMapper.toCartItemDTO(savedCartItem);
         }
         return null;
     }
 
     @Override
     public void deleteCartItem(CartItemKey id) {
-        cartItemRepository.deleteById(id);
+        Optional<CartItem> cartItem = cartItemRepository.findById(id);
+        if (cartItem.isPresent()) {
+            cartItemRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Cart item not found with id: " + id);
+        }
     }
-
-    @Override
-    public List<CartItem> getCartItemByPage(List<CartItem>cartItems,int page, int pageSize) {
-        int total = cartItems.size();
-        int fromIndex = Math.min((page - 1) * pageSize, total);
-        int toIndex = Math.min(page * pageSize, total);
-        return cartItems.subList(fromIndex, toIndex);
-    }
-
-
 }
-
-
